adding the whole file

+++ src/main/java/com/zimpy/user/UserController/AddressController.java
@@ -1,5 +1,6 @@
 package com.zimpy.user.UserController;
 
+import com.zimpy.exception.UnauthorizedException;
 import com.zimpy.user.Services.AddressService;
 import com.zimpy.user.dto.AddressRequest;
 import com.zimpy.user.dto.AddressResponse;
@@ -17,32 +18,47 @@ public class AddressController {
 
     private final AddressService addressService;
 
-    public AddressController(AddressService addressService){
+    public AddressController(AddressService addressService) {
         this.addressService = addressService;
     }
 
     @PostMapping
-    public ResponseEntity<AddressResponse> addAddress(Authentication authentication , @Valid @RequestBody AddressRequest request){
+    public ResponseEntity<AddressResponse> addAddress(Authentication authentication,
+            @Valid @RequestBody AddressRequest request) {
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new UnauthorizedException("please login first");
+        }
         Long userId = (Long) authentication.getPrincipal();
-        return ResponseEntity.ok(addressService.addAddress(userId,request));
+        return ResponseEntity.ok(addressService.addAddress(userId, request));
     }
+
     @GetMapping
-    public ResponseEntity<List<AddressResponse>> myAddress(Authentication authentication){
+    public ResponseEntity<List<AddressResponse>> myAddress(Authentication authentication) {
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new UnauthorizedException("please login first");
+        }
         Long userId = (Long) authentication.getPrincipal();
         return ResponseEntity.ok(addressService.getMyAddresses(userId));
     }
+
     @PutMapping("/{id}")
     public ResponseEntity<AddressResponse> updateAddress(
+
             @PathVariable Long id,
             @Valid @RequestBody AddressRequest request,
-            Authentication authentication
-    ) {
+            Authentication authentication) {
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new UnauthorizedException("please login first");
+        }
         Long userId = (Long) authentication.getPrincipal();
         return ResponseEntity.ok(addressService.updateAddress(userId, id, request));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String >deleteAddress(@PathVariable Long id, Authentication authentication){
+    public ResponseEntity<String> deleteAddress(@PathVariable Long id, Authentication authentication) {
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new UnauthorizedException("please login first");
+        }
         Long userId = (Long) authentication.getPrincipal();
         return ResponseEntity.ok(addressService.deleteAddress(userId, id));
     }
